controllers/version/metadata: Move `get_version_authors()` to dedicated module
